requirements.txt
# All Python dependencies
numpy
opencv-python-headless
Pillow
PyYAML
watchdog
pyftpdlib
PyMuPDF
reportlab
onnxruntime
withoutbg
deskew

# Web UI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
